requirements.txt
django-allauth==0.63.3
django-phonenumber-field>=7.0.0
django-redis==5.4.0
django-rq==2.10.2
djangorestframework==3.14.0
psycopg2-binary==2.9.9
python-dotenv==0.21.1
phonenumbers==8.13.39
requests==2.22.0
PyJWT==2.3.0
cryptography==3.4.7
gunicorn==22.0.0
dj-database-url==2.2.0
whitenoise==6.7.0
uvicorn==0.30.1
